Create utils/ids utility module
